Made Abstract Batbox class
