+++ src/main/java/baboon/industry/block/storage/entity/TileEntityTransformerBase.java
@@ -2,8 +2,8 @@ package baboon.industry.block.storage.entity;
 
 import baboon.industry.block.entity.TileEntityEnergyConductorDamageable;
 import com.mojang.nbt.CompoundTag;
-import sunsetsatellite.sunsetutils.util.Connection;
-import sunsetsatellite.sunsetutils.util.Direction;
+import sunsetsatellite.catalyst.core.util.Connection;
+import sunsetsatellite.catalyst.core.util.Direction;
 
 public class TileEntityTransformerBase extends TileEntityEnergyConductorDamageable {
